% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamour-package.R
\docType{package}
\name{clamour-package}
\alias{clamour}
\alias{clamour-package}
\title{clamour: Website for Twitter analysis results}
\description{
clamour provides a template for setting up a website to display the
    results of analysis of Twitter hashtags.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lazappi/clamour}
  \item Report bugs at \url{https://github.com/lazappi/clamour/issues}
}

}
\author{
\strong{Maintainer}: Luke Zappia \email{luke@lazappi.id.au} (0000-0001-7744-8565)

}
\keyword{internal}
