% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{clamour_list}
\alias{clamour_list}
\title{clamour list}
\usage{
clamour_list(path = ".", ignore = c("index.Rmd", "about.Rmd",
  "_analysis.Rmd"))
}
\arguments{
\item{path}{Path to directory to search for \code{.Rmd} files}

\item{ignore}{Vector or files to ignore}
}
\value{
Markdown string describing analysis files
}
\description{
Produce a Markdown string with details of all the clamour analysis files in a
directory. This function is primarily designed to be used in the index of a
website so the list is automatically updated.
}
