% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{clamour_new}
\alias{clamour_new}
\title{New clamour analysis}
\usage{
clamour_new(
  name,
  hashtag,
  description = paste("An analysis of", hashtag),
  start_day = Sys.Date(),
  end_day = Sys.Date() + 1,
  timezone = Sys.timezone(),
  theme = "theme_light",
  accent = "dodgerblue",
  accent2 = NULL,
  kcore = 2,
  topics_k = 6,
  bigram_filter = 3,
  fixed = FALSE,
  seed = 1,
  dir = "analysis"
)
}
\arguments{
\item{name}{Name of the analysis. This is used to name files so should
usually be short and simple.}

\item{hashtag}{The hashtag to use for this analysis, starting with "#". Can
also be a vector of hashtags.}

\item{description}{A more detailed description of the analysis, details of
the event etc.}

\item{start_day}{First day to be included in the analysis.}

\item{end_day}{Last day to be included in the analysis.}

\item{timezone}{Timezone to be used for the analysis.}

\item{theme}{Name of a \href{https://ggplot2.tidyverse.org/}{ggplot2} theme
function to use for plots.}

\item{accent}{Accent colour to use for plots.}

\item{accent2}{Secondary accent colour to use for plots. If \code{NULL} this will
be created based on \code{accent}.}

\item{kcore}{k-core to use for filtering the mentions network.}

\item{topics_k}{Number of topics to use for topic modelling.}

\item{bigram_filter}{Number of co-occurrences for filtering the bigram graph.}

\item{fixed}{Whether to continue to search and download new tweets. Set to
\code{TRUE} to fix the dataset used for analysis.}

\item{seed}{Random seed set at the start of analysis.}

\item{dir}{Directory to save the new analysis file.}
}
\value{
Path to the newly created file, invisibly
}
\description{
Creates a new analysis file based on the clamour template.
}
