% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{clamour_table}
\alias{clamour_table}
\title{clamour table}
\usage{
clamour_table(path = ".", ignore = c("index.Rmd", "about.Rmd",
  "_analysis.Rmd"))
}
\arguments{
\item{path}{Path to directory to search for \code{.Rmd} files}

\item{ignore}{Vector or files to ignore}
}
\value{
data.frame containing information on analysis files
}
\description{
Read the YAML front matter from a set of clamour analysis files and produce
a tidy data.frame.
}
